Add section on past changes, as old files might still linger
In the past there has been a change in how variable blocksize
files are coded. There are still FLAC files and encoder around
that follow the old format. For example, the current flake
package in Debian repositories still create such files.

Therefore, it seemed fitting to add a section about what
changed, so decoders might be able to recognize such files
if necessary.