Add appendix with example file decoding 'by hand'
In a previous PR, I added an example to the coded residual section
which felt out of place. This examples can provide a more hands-on
way to understand the FLAC specification for readers that need it.
Also, it can be used by people proofreading this specification to
do cross-referencing, as it is 'redundant'.