Add reserved bit to synccode (#161)
In the original specification, the synccode at the start of an
audio frame was 14 bit, with a note that the bit following the
sync code, which was 'reserved' must remain 0 in order to
distinguish a FLAC frame from an MPEG frame.

Details:
https://lists.xiph.org/pipermail/flac-dev/2008-November/002598.html
https://lists.xiph.org/pipermail/flac-dev/2008-December/002607.html
https://lists.xiph.org/pipermail/flac-dev/2009-January/002622.html
https://github.com/xiph/flac/commit/bd161c61fa665fb4e91485e2107946f73affcdb8

In previous commits I already mentioned a 15 bit sync code, but
this wasn't consistent. This commit changes a mention of a 14-bit
sync code to a 15-bit one. Also, it improves the wording of
including the blocksize strategy bit in the search for frame sync