At appendix on past format changes
In the past there has been a change in how variable blocksize
files are coded. There are still FLAC files and encoders around
that follow the old format. For example, the current flake package
in Debian repositories still create such files.

Therefore, it seemed fitting to add an appendix about what
changed, so decoders might be able to recognize such files
if necessary.

Also, the changes to the format by CELLAR are documented